+++ GameEngine/Include/IMGUIGrid.cpp
@@ -34,10 +34,10 @@ void CIMGUIGrid::Render()
 
 	Resolution Res = CDevice::GetInst()->GetResolution();
 	Vector2 Ratio = CDevice::GetInst()->GetViewportAspectRatio();
-	Res.Width *= Ratio.x;
-	Res.Height *= Ratio.y;
+	Res.Width *= (unsigned int)Ratio.x;
+	Res.Height *= (unsigned int)Ratio.y;
 
 	ImGuiIO& io = ImGui::GetIO();
-	ImGuizmo::SetRect(Rect.left, Rect.top,Res.Width, Res.Height);
+	ImGuizmo::SetRect((float)Rect.left, (float)Rect.top,(float)Res.Width, (float)Res.Height);
 	ImGuizmo::DrawGrid((float*)&matView, (float*)&matProj, (float*)&m_matIdentity, m_Size.x);
 }
